backend/requirements.txt
# This file specifies the third-party dependencies for the SynapseERP project.
# It uses flexible versioning (>=) to ensure portability across Python versions.
# Note: the editable install (-e backend/) is handled by run.sh, not listed here.

# Core Framework
Django>=5.0,<6.0
djangorestframework>=3.15
gunicorn

# Data Analysis
pandas>=2.2
numpy>=1.26.4
openpyxl

# Utilities
python-dotenv
toml
requests
PyYAML>=6.0
markdown2
colorlog
polib
googletrans==4.0.0rc1 # Pinned because it is a release candidate

# Indirect dependencies (usually handled by pip, listed for clarity if needed)
asgiref
certifi
chardet
charset-normalizer
et_xmlfile
h11
h2
hpack
hstspreload
httpcore<1.0
httpx<0.28
hyperframe
idna
packaging
python-dateutil
pytz
rfc3986
setuptools
six
sniffio
sqlparse
tzdata
urllib3
wheel
